requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
crispy-bootstrap5==0.7
Django
django-crispy-forms==2.0
django-gravatar2==1.4.4
django-jazzmin==2.6.0
django-model-utils==4.3.1
django-modeltranslation==0.18.11
django-suit @ https://github.com/darklow/django-suit/tarball/v2#sha256=b40c916de9528ca86bcccd42d1fd54e7773e5071d2bfd9efc576039c1a441c0b
django-widget-tweaks==1.4.12
idna==3.4
isodate==0.6.1
lxml==4.9.3
Pillow==10.0.0
platformdirs==3.10.0
pycryptodome==3.18.0
psycopg2==2.9.9
pytz==2023.3
requests==2.31.0
requests-file==1.5.1
requests-toolbelt==1.0.0
six==1.16.0
sqlparse==0.4.4
swapper==1.3.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
zeep==4.2.1
